% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpx_ms_age.R
\name{age.distribution}
\alias{age.distribution}
\title{Calculate age distribution in equilibrium}
\usage{
age.distribution(contact.matrix, R0, vaccination = NULL, years = NULL)
}
\arguments{
\item{R0}{The value(s) of R0. If a vactor of length>1, will test
the different values}

\item{vaccination}{Vaccination levels, given as a data table with
a "year" column, and a column for each of the age groups}

\item{years}{Years for which to calculate age distribution. This
takes into account vaccination in the different years. If this is
given, R0 should be set to a single value}

\item{agegroups}{Vector of the lower limits of the age groups}
}
\value{
a data.table with the different relative incidences,
population sizes, absolute incidences and proportions of cases
}
\description{
Calculates the age distribution in equilibrium of an age-strutured
SIR-model, by year or as a function of R0
}
\author{
Sebastian Funk
}

