% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpx_ms_age.R
\name{age.likelihood}
\alias{age.likelihood}
\title{Log-likelihood for a model run of an age-structured
childhood infection model}
\usage{
age.likelihood(trajectory, data.cases, data.sero = NULL, rel = F,
  sample = NULL, proportion.reported = 1, normal.approximation = F,
  time.column = "time")
}
\arguments{
\item{trajectory}{The model trajectory, in a data.table with a
column "year", a column "lower.age.limit", and at least one of
"rel.incidence" or "abs.incidence"}

\item{data.cases}{The data trajectory, in the same format as the model trajectory}

\item{data.sero}{Serology, if any}

\item{rel}{Whether to fit relative incidences}

\item{sample}{Whether the data are the result of a sampling procedure}

\item{proportion.reported}{The average proportion of cases reported}

\item{normal.approximation}{Whether to use the normal
approximation to the binomial distribution}
}
\value{
A number, the log-likelihood
}
\description{
Log-likelihood for a model run of an age-structured
childhood infection model
}
\author{
Sebastian Funk
}

