% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{agegroups.to.limits}
\alias{agegroups.to.limits}
\title{Convert age groups to lower age limits.}
\usage{
agegroups.to.limits(groups)
}
\arguments{
\item{groups}{age groups}
}
\value{
lower age limits
}
\description{
Mostly used to go from a contact matrix to lower limits
}

