% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpx_ms_age.R
\name{backcalc.population}
\alias{backcalc.population}
\title{Calculate previous states of a population, using a demographic model}
\usage{
backcalc.population(nyears = 1, lower.age.limits = NULL)
}
\arguments{
\item{nyears}{number of years to backcalculate}

\item{lower.age.limits}{lower limits of age groups}
}
\value{
named vector of states
}
\description{
Calculate previous states of a population, using a demographic model
}
\author{
Sebastian Funk
}

