% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/libbi.R
\name{bi_string}
\alias{bi_string}
\title{Generates a bi string from a string and a vector}
\usage{
bi_string(l)
}
\arguments{
\item{l}{list with a 'string' (character) and 'argument- (vector or list)}
}
\value{
bi string
}
\description{
Internal function used bi generate_bi
}
\author{
Sebastian Funk
}

