% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/libbi.R
\name{bi_write}
\alias{bi_write}
\title{Writes a vector to a file connector (and to the screen, if descired)}
\usage{
bi_write(text, con = stdout(), print = FALSE, ...)
}
\arguments{
\item{text}{text to print}

\item{con}{file connector (stdout by default)}

\item{print}{whether to also print the text to screen (ignored if
\code{con} is set to stdout)}
}
\value{
result of the write to the file connector
}
\description{
Internal function used bi generate_bi
}
\author{
Sebastian Funk
}

