% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cpx_ms_age.R
\name{brownian}
\alias{brownian}
\title{Generate a trajectory of brownian motion}
\usage{
brownian(parameters, n)
}
\arguments{
\item{parameters}{A named vector or list that includes the initial
value ("init"), drift ("drift") and scale ("scale") as named
parameters}

\item{n}{Number of steps to compute}
}
\value{
Brownian trajectory
}
\description{
Generate a trajectory of brownian motion, given an initial value,
as well as drift and scale parameters
}
\author{
Sebastian Funk
}

