% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{clean_countries}
\alias{clean_countries}
\title{Clean country names}
\usage{
clean_countries(table, column = "country")
}
\arguments{
\item{table}{data table which contains country names}

\item{column}{the column with country names}
}
\value{
cleaned data table
}
\description{
Clean country names
}
\author{
Sebastian Funk
}

