% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{combine.compartments}
\alias{combine.compartments}
\title{Combine compartments that have been split up to get an Erlang distribution}
\usage{
combine.compartments(traj)
}
\arguments{
\item{traj}{trajectory data frame}
}
\value{
trajectory data frame with Erlang compartments summed up
}
\description{
Combine compartments that have been split up to get an Erlang distribution
}
\author{
Sebastian Funk
}

