% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpx_ms_age.R
\name{cpx.age.wallinga}
\alias{cpx.age.wallinga}
\title{Chickenpox age distribution of cases from the Wallinga (2004) method}
\usage{
cpx.age.wallinga(...)
}
\arguments{
\item{...}{parameters to be passed to fit.polymod}
}
\value{
the results of fit.polymod
}
\description{
Chickenpox age distribution of cases from the Wallinga (2004) method
}
\author{
Sebastian Funk
}

