% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpx_ms_age.R
\name{cpx.init.prior}
\alias{cpx.init.prior}
\title{Chickenpox prior for initial conditions}
\usage{
cpx.init.prior(init)
}
\arguments{
\item{init}{initial conditions}
}
\value{
(logged) prior
}
\description{
Chickenpox prior for initial conditions
}
\author{
Sebastian Funk
}

