% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cpx_ms_age.R
\name{cpx.init}
\alias{cpx.init}
\title{Generate initial conditions for chickenpox from given proportion
of susceptibles}
\usage{
cpx.init(parameters)
}
\arguments{
\item{parameters}{Model parameters}
}
\value{
Named vector of initial conditions
}
\description{
Generate initial conditions for chickenpox from given proportion
of susceptibles
}
\author{
Sebastian Funk
}

