% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/time_series.R
\name{df_to_ts}
\alias{df_to_ts}
\title{Convert a \code{data.frame} into a \code{ts} time series object}
\usage{
df_to_ts(data, interval = "day", date.column = NULL, value.column = NULL,
  start = NULL, end = NULL, ...)
}
\arguments{
\item{data}{data to convert}

\item{interval}{data interval (day, week, month, quarter, year, or an arbitrary number)}

\item{date.column}{date column; if not given will be guessed as the first column of class \code{Date}; if not given and no column is of class \code{Date}, will quit with an error}

\item{value.column}{value column; if not given, will assume that dates are individual data per date (which will be added up)}

\item{start}{starting date}

\item{end}{end date}

\item{...}{}
}
\value{
A time series (\code{ts}) object
}
\description{
Convert a \code{data.frame} into a \code{ts} time series object
}
\author{
Sebastian Funk
}

