% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallinga.R
\name{endemic.age.dist}
\alias{endemic.age.dist}
\title{calculates the age distribution in an endemic setting using the iterative method
of Wallinga (2006)}
\usage{
endemic.age.dist(contact.matrix, q, child.mixing = NA, reporting = NA,
  vaccination = NA, incidence.start = 0.01, tol = 1e-05)
}
\arguments{
\item{contact.matrix:}{social contact matrix; this must have the
age groups as column names (with a comma between limits)}

\item{q:}{vector of probabilies of transmission per contact}

\item{child.mixing:}{vector of first entries of contact matrix}

\item{reporting:}{vector of reporting probabilities}

\item{vaccination:}{proportion vaccinated (between 0 and 1) in each age group}

\item{incidence.start:}{starting value for inidence}

\item{tol:}{tolerance for stopping the iteration}
}
\value{
the endemic age distribution, a list composed of:
  y: incidence in each age group
  normalised.y: proportion of cases in each age group
  lambda: force of infection by age group
  prop.immune: proportion immune in each age group
  prop.immune.age: proportion immune at each age
}
\description{
calculates the age distribution in an endemic setting using the iterative method
of Wallinga (2006)
}

