% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallinga.R
\name{epidemic.age.dist}
\alias{epidemic.age.dist}
\title{calculates the age distribution in an epidemic setting using the iterative method
of Wallinga (2006)}
\usage{
epidemic.age.dist(contact.matrix, q, child.mixing = NA, immunity = 0,
  final.size.start = 0.01, tol = 0.00001)
}
\arguments{
\item{contact.matrix:}{social contact matrix}

\item{age.sizes:}{sizes of (number of people in) the different age groups}

\item{q:}{probability of transmission per contact}

\item{child.mixing:}{first entry of contact matrix}

\item{immunity:}{proportion immune before the epidemic}

\item{incidence.start:}{starting value for inidence}

\item{tol:}{tolerance for stopping the iteration}
}
\value{
A list composed of
  z: final size in each age group
  normalised.z: proportion of cases in each age group
}
\description{
calculates the age distribution in an epidemic setting using the iterative method
of Wallinga (2006)
}

