% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpx_ms_age.R
\name{estimate.immunity.mmr}
\alias{estimate.immunity.mmr}
\title{Estimate vaccine-induced immunity in the population from reported MMR coverage}
\usage{
estimate.immunity.mmr(vaccine.efficacy = 0.9, newborn.immunity = NULL,
  years = NULL, age.limits = NULL)
}
\arguments{
\item{vaccine.efficacy}{Efficacy of one dose of the vaccien}

\item{newborn.immunity}{Immunity of newborns for hte first year of
their lives. If not given, this is extracted from serology data}

\item{years}{the years for which to calculate immunity}
}
\value{
A data table with vaccine-induced immunity in all age groups
}
\description{
Estimate vaccine-induced immunity in the population from reported MMR coverage
}
\author{
Sebastian Funk
}

