% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cpx_ms_age.R
\name{estimate.uptake.mmr}
\alias{estimate.uptake.mmr}
\title{Estimate vaccine uptake by age group in the population from reported MMR coverage}
\usage{
estimate.uptake.mmr()
}
\arguments{
\item{vaccine.efficacy}{Efficacy of one dose of the vaccien}

\item{years}{the years for which to calculate immunity}
}
\value{
A data table with vaccine-induced immunity in all age groups
}
\description{
Estimate vaccine uptake by age group in the population from reported MMR coverage
}
\author{
Sebastian Funk
}

