% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cpx_ms_age.R
\name{ew.pop.parameters}
\alias{ew.pop.parameters}
\title{Get population parameters for England & Wales}
\usage{
ew.pop.parameters(age.limits, year.limits = NULL, interpolate = FALSE)
}
\arguments{
\item{age.limits}{Lower limits of the age groups}

\item{year.limits}{first and last year}

\item{extra.parameters}{Extra parameters, especially whether we are interpolating (in which case we ##' need another year of data at both ends)}
}
\value{
list of parameters
}
\description{
Fill parameters for births, deaths and ageing, with England &
Wales population data
}
\author{
Sebastian Funk
}

