% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cpx_ms_age.R
\name{fit.measles.wallinga}
\alias{fit.measles.wallinga}
\title{Fit the measles data}
\usage{
fit.measles.wallinga(years = NULL, contact.matrix = NULL)
}
\arguments{
\item{years}{The years to fit (of not given, fit all years)}

\item{contact.matrix}{The contact matrix to use (if not given,
sample one from UK Polymod)}
}
\value{
A list with the best fits (q, rep and child.mixing), the
used contact.matrix, and the return of fitting q and rep with any
messages and convergence information that may have been reported.
}
\description{
Fit a model with a free contact rate between <5 year olds to the
England & Wales measles data, assuming everything is in
equilibrium every year (a method similar to Wallinga, Teunis and
Kretzschmar, 2006, "Using Data on Social Contacts to Estimate
Age-specific Transmission Parameters for Respiratory-spread
Infectious Agents", Am J Epidemiol, 164, 936-944. This estimates
q, the reporting rate, and the mixing between children every year.
}

