% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polymod.R
\name{fit.polymod}
\alias{fit.polymod}
\title{Fit POLYMOD matrix and age population to target distribution of cases}
\usage{
fit.polymod(participants, contacts, ages, target, nFits = 1, sample = F,
  yearbyyear = F, verbose = F, ...)
}
\arguments{
\item{participants}{POLYMOD participants}

\item{contacts}{POLIYMOD contacts}

\item{ages}{age groups and number of people in each of these
groups. This is needed to normalise the contact matrix}

\item{target}{target relative age distribution}

\item{nFits}{nubmer of fits to produce}

\item{sample}{whether to take a sample from the age of contacts}

\item{yearbyyear}{whether to fit q year-by-year instead of fitting one parameter}
}
\value{
a list containing
  q: best fitting q
  child.mixing: best fixing mixing in first age group
  R0: R0 for the best fitting parameters
  diff: difference between the fitted and target age distributions
}
\description{
Takes a polymod matrix (participants and contacts) and a distribution of ages in
the population and fits it to a target distribution of cases among ages (fitting
the transmission parameter q and mixing between the first age group)
}
\details{
This is specific to the POLYMOD data format
}

