% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cpx_ms_age.R
\name{forward.initial.conditions}
\alias{forward.initial.conditions}
\title{Update initial conditions to the start of the actual run (not run-in)}
\usage{
forward.initial.conditions(parameters, data, gen.init, thickening = 0,
  time.column = "time")
}
\arguments{
\item{parameters}{model parameters}

\item{data}{data (needed for run times and age limits etc)}

\item{gen.init}{function to generate initial conditions}

\item{thickening}{whether to thicken the runs}

\item{time.column}{time column in the data}
}
\value{
updated initial conditions
}
\description{
Update initial conditions to the start of the actual run (not run-in)
}
\author{
Sebastian Funk
}

