% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpx_ms_age.R
\name{gen.init}
\alias{gen.init}
\title{Generate initial conditions for measles or chickenpox from given proportion
of susceptibles}
\usage{
gen.init(parameters, infection = c("measles", "chickenpox"))
}
\arguments{
\item{parameters}{Model parameters}

\item{infection}{"measles" or "chickenpox"}
}
\value{
Named vector of initial conditions
}
\description{
Generate initial conditions for measles or chickenpox from given proportion
of susceptibles
}
\author{
Sebastian Funk
}

