% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/libbi.R
\name{generate_bi}
\alias{generate_bi}
\title{Write a bi file for an SEIR model}
\usage{
generate_bi(name = "SEIR", model, desc = NULL, print = FALSE)
}
\arguments{
\item{name}{name of the model (and file)}

\item{model}{list of things -- see code}

\item{desc}{description of the model}
}
\value{
nothing -- a libbi file is written (unless \code{print} is
set to TRUE, in which case the output is also printed to screen
}
\description{
Write a bi file for an SEIR model
}
\author{
Sebastian Funk
}

