% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/libbi.R
\name{generate_bi_seir}
\alias{generate_bi_seir}
\title{Write a bi file for an SEIR model}
\usage{
generate_bi_seir(name = "SEIR", model, desc = NULL, print = FALSE)
}
\arguments{
\item{model}{list of things -- see code}

\item{groups}{groups (if there are any)}

\item{delta}{delta for the transition block}

\item{sum.agegroups}{whether to sum up age groups for incidence}

\item{...}{parameters to pass to generate_bi}
}
\value{
nothing -- a libbi file is written
}
\description{
Write a bi file for an SEIR model
}
\author{
Sebastian Funk
}

