% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polymod.R
\name{homogeneous.mixing.matrix}
\alias{homogeneous.mixing.matrix}
\title{Generate a homogeneous mixing matrix from a social contact matrix.}
\usage{
homogeneous.mixing.matrix(contact.matrix)
}
\arguments{
\item{contact.matrix}{the matrix to homogenise}
}
\value{
homogeneous mixing matrix
}
\description{
Generates a homogeneous mixing matrix with the same age groups as
a given contact matrix
}
\author{
Sebastian Funk
}

