% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seir.R
\name{initSIR}
\alias{initSIR}
\title{Initial conditions for the SIR model}
\usage{
initSIR(state, parameters, additional = c("Z"))
}
\arguments{
\item{state}{named state vector}

\item{parameters}{parameters of the model}

\item{additional}{additional classes}
}
\value{
a named vector of initial conditions
}
\description{
Takes a vector of initial conditions (in S, E(optionally), I, R)
and returns a vector of initial conditions to use with S(E)IR
models of arbitrary number of infectious stages and agegroups
}
\details{
Multiple exposed or infectious stages might be introduced to
change the distribution of waiting times, or because there are
biologically distinct stages. Multiple age groups might be
introduced to follow infection across age groups.

The number of age groups is determined by the number of S classes
in the state vector
}
\author{
Sebastian Funk
}

