% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpx_ms_age.R
\name{interpolate.serology}
\alias{interpolate.serology}
\title{Interpolate serology}
\usage{
interpolate.serology(age.limits)
}
\arguments{
\item{age.limits}{The lower limits of the required age classes}
}
\value{
interpolated serologies
}
\description{
This splits age classes into other age classes by assuming that
results were randomly distributed between ages
}
\author{
Sebastian Funk
}

