% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/libbi.R
\name{libbi.get.likelihoods}
\alias{libbi.get.likelihoods}
\title{Get likelihoods from a libbi results file}
\usage{
libbi.get.likelihoods(filename, initial = F)
}
\arguments{
\item{filename}{name of the libbi results file, in ncdf format}

\item{initial}{whether initial conditions were also fitted}
}
\value{
a data frame of log-likelihoods and corresnponding parameters
}
\description{
Get likelihoods from a libbi results file
}
\details{
Get likelihoods and associated parameter values from a libbi results file
}
\author{
Sebastian Funk
}

