% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/libbi.R
\name{libbi.get.params}
\alias{libbi.get.params}
\title{Get parameter values from libbi results file}
\usage{
libbi.get.params(filename)
}
\arguments{
\item{filename}{name of the libbi results file (in ncdf format)}
}
\value{
data table of parameter values
}
\description{
Get parameter values from libbi results file
}
\details{
Loop over all one-dimensional entries of the supplied ncdf file
(except the first variable, which is time) and store them in a
data table
}
\author{
Sebastian Funk
}

