% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/libbi.R
\name{libbi.get.trajectory}
\alias{libbi.get.trajectory}
\title{Get a fit from libbi results file}
\usage{
libbi.get.trajectory(filename, row = NA_integer_)
}
\arguments{
\item{filename}{name of the libbi results file (in ncdf format)}

\item{row}{if not NA, the number of the fit one wants to get}
}
\value{
a list of variables, with a trajectory for each
}
\description{
Get a trajectory from libbi results file
}
\details{
Gets the trajectory with the lowest largest log-likelihood or, if a row
is specified, a specific fit from a libbi results file
}
\author{
Sebastian Funk
}

