% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{limits.to.agegroups}
\alias{limits.to.agegroups}
\title{Convert lower age limits to age groups.}
\usage{
limits.to.agegroups(limits)
}
\arguments{
\item{limits}{Lower age limits}
}
\value{
Age groups (limits separated by dashes)
}
\description{
Mostly used for plot labelling
}

