% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_data.R
\name{loadBryanData}
\alias{loadBryanData}
\title{Load measles England & Wales data}
\usage{
loadBryanData(path)
}
\arguments{
\item{path}{Path in which to find the raw data}
}
\description{
Load measles England & Wales data
}
\author{
Sebastian Funk
}

