% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_data.R
\name{loadChickenpoxEWAge}
\alias{loadChickenpoxEWAge}
\title{Load Chickenpox England & Wales age data}
\usage{
loadChickenpoxEWAge(path)
}
\arguments{
\item{path}{Path in which to find the raw data}
}
\description{
Load Chickenpox England & Wales age data
}
\author{
Sebastian Funk
}

