% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{loadChickenpoxSerology}
\alias{loadChickenpoxSerology}
\title{Load Chickenpox serology data}
\usage{
loadChickenpoxSerology(path)

loadChickenpoxSerology(path)
}
\arguments{
\item{path}{Path in which to find the raw data}

\item{path}{Path in which to find the raw data}
}
\description{
Load Chickenpox serology data

Load Childcare data
}
\author{
Sebastian Funk

Sebastian Funk
}

