% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_data.R
\name{loadCopenhagenDisease}
\alias{loadCopenhagenDisease}
\title{Load Copenhagen disease data}
\usage{
loadCopenhagenDisease(path)
}
\arguments{
\item{path}{Path in which to find the raw data}
}
\description{
Load Copenhagen disease data
}
\author{
Sebastian Funk
}

