% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{loadDemographicsEW}
\alias{loadDemographicsEW}
\title{Load demographic England & Wales data}
\usage{
loadDemographicsEW(path)
}
\arguments{
\item{path}{Path in which to find the raw data}
}
\description{
Load demographic England & Wales data
}
\author{
Sebastian Funk
}

