% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_data.R
\name{loadEuropeanMaps}
\alias{loadEuropeanMaps}
\title{Load European maps}
\usage{
loadEuropeanMaps(shapePath)
}
\arguments{
\item{path}{Path in which to find the raw data}
}
\description{
Load European maps
}
\author{
Sebastian Funk
}

