% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{loadMeaslesEurope}
\alias{loadMeaslesEurope}
\title{Load measles Europe data}
\usage{
loadMeaslesEurope(path)
}
\arguments{
\item{path}{Path in which to find the raw data}
}
\description{
Load measles Europe data
}
\author{
Sebastian Funk
}

