% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_data.R
\name{loadMeaslesEWConfirmation}
\alias{loadMeaslesEWConfirmation}
\title{Load Measles England & Wales confirmation data}
\usage{
loadMeaslesEWConfirmation(path)
}
\arguments{
\item{path}{Path in which to find the raw data}
}
\description{
Load Measles England & Wales confirmation data
}
\author{
Sebastian Funk
}

