% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{loadMeaslesSerology}
\alias{loadMeaslesSerology}
\title{Load measles serology data}
\usage{
loadMeaslesSerology(path)
}
\arguments{
\item{path}{Path in which to find the raw data}
}
\description{
Load measles serology data
}
\author{
Sebastian Funk
}

