% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{loadMeaslesSerologyFine1950}
\alias{loadMeaslesSerologyFine1950}
\title{Load measles serology data from Fine(1950)}
\usage{
loadMeaslesSerologyFine1950(path)
}
\arguments{
\item{path}{Path in which to find the raw data}
}
\description{
Load measles serology data from Fine(1950)
}
\author{
Sebastian Funk
}

