% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_data.R
\name{loadMeaslesWHO}
\alias{loadMeaslesWHO}
\title{Load measles WHO data}
\usage{
loadMeaslesWHO(path)
}
\arguments{
\item{path}{Path in which to find the raw data}
}
\description{
Load measles WHO data
}
\author{
Sebastian Funk
}

