% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{loadMMREW}
\alias{loadMMREW}
\title{Load MMR England & Wales data}
\usage{
loadMMREW(path)
}
\arguments{
\item{path}{Path in which to find the raw data}
}
\description{
Load MMR England & Wales data
}
\author{
Sebastian Funk
}

