% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{loadMMRWorld}
\alias{loadMMRWorld}
\title{Load MMR world data}
\usage{
loadMMRWorld(path)
}
\arguments{
\item{path}{Path in which to find the raw data}
}
\description{
Load MMR world data
}
\author{
Sebastian Funk
}

