% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{loadPolymod}
\alias{loadPolymod}
\title{Load POLYMOD data}
\usage{
loadPolymod(path)
}
\arguments{
\item{path}{Path in which to find the raw data}
}
\description{
Load POLYMOD data
}
\author{
Sebastian Funk
}

