% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_data.R
\name{loadPopulationEurope}
\alias{loadPopulationEurope}
\title{Load population Europe data}
\usage{
loadPopulationEurope(path)
}
\arguments{
\item{path}{Path in which to find the raw data}
}
\description{
Load population Europe data
}
\author{
Sebastian Funk
}

