% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{loadPopulationEWAge}
\alias{loadPopulationEWAge}
\title{Load population England & Wales age data}
\usage{
loadPopulationEWAge(path, mult.factor = 1000)
}
\arguments{
\item{path}{Path in which to find the raw data}
}
\description{
Load population England & Wales age data
}
\author{
Sebastian Funk
}

