% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{loadPopulationWorldAge}
\alias{loadPopulationWorldAge}
\title{Load population world data}
\usage{
loadPopulationWorldAge(path)
}
\arguments{
\item{path}{Path in which to find the raw data}
}
\description{
Load population world data
}
\author{
Sebastian Funk
}

