% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_data.R
\name{loadRVCVaccinationData}
\alias{loadRVCVaccinationData}
\title{Load SIA data}
\usage{
loadRVCVaccinationData(path)
}
\arguments{
\item{path}{Path in which to find the raw data}
}
\description{
Load SIA data
}
\author{
Sebastian Funk
}

