% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{loadWorldMeaslesData}
\alias{loadWorldMeaslesData}
\title{Load measles world data}
\usage{
loadWorldMeaslesData(path)
}
\arguments{
\item{path}{Path in which to find the raw data}
}
\description{
Load measles world data
}
\author{
Sebastian Funk
}

