% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{loadWorldPopulationProspectData}
\alias{loadWorldPopulationProspectData}
\title{Load worldwide serology data}
\usage{
loadWorldPopulationProspectData(path)
}
\arguments{
\item{path}{Path in which to find the raw data}
}
\description{
Load worldwide serology data
}
\author{
Sebastian Funk
}

