% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{lower.to.upper.limits}
\alias{lower.to.upper.limits}
\title{Convert lower to upper age limits}
\usage{
lower.to.upper.limits(x, limits, max = 100)
}
\arguments{
\item{x}{the age group to convert}

\item{limits}{overall lower age limits}

\item{max}{maximum age}
}
\value{
upper age limits
}
\description{
Convert lower to upper age limits
}
\author{
Sebastian Funk
}

