% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/libbi.R
\name{make.ncdf}
\alias{make.ncdf}
\title{Make ncdf file (for libbi) from data}
\usage{
make.ncdf(data, filename, date.var = "rash.date", value.var = "cases",
  data.var = NA_character_, fill.date = NA_integer_, start.date = NULL,
  t0 = 0, delta = 1)
}
\arguments{
\item{data}{dataset to convert}

\item{filename}{filename}

\item{date.var}{date variable}

\item{value.var}{value variable}

\item{data.var}{what to call the data}

\item{fill.date}{whether to fill the data with zeroes}

\item{t0}{time 0}

\item{start}{time at which to start looking at the data}
}
\value{
the outcome of \code{close.ncdf}
}
\description{
Make ncdf file (for libbi) from data
}
\author{
Sebastian Funk
}

