% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{melt.trajectory}
\alias{melt.trajectory}
\title{Melt a trajectory in terms of time and age groups}
\usage{
melt.trajectory(traj, age.labels = NULL, time.label = NULL)
}
\arguments{
\item{traj}{Trajectory}

\item{age.labels}{labels of the age groups (optional)}
}
\value{
a nice, melted trajectory
}
\description{
Melt a trajectory in terms of time and age groups
}
\author{
Sebastian Funk
}

