% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpx_ms_age.R
\name{ms.age.structure.wallinga}
\alias{ms.age.structure.wallinga}
\title{Measles age distribution of cases from the Wallinga (2004) method}
\usage{
ms.age.structure.wallinga(...)
}
\arguments{
\item{...}{parameters to be passed to fit.polymod}
}
\value{
the results of fit.polymod
}
\description{
Measles age distribution of cases from the Wallinga (2004) method
}
\author{
Sebastian Funk
}

