% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cpx_ms_age.R
\name{ms.init}
\alias{ms.init}
\title{Generate initial conditions for measles from given proportion
of susceptibles}
\usage{
ms.init(parameters)
}
\arguments{
\item{parameters}{Model parameters}
}
\value{
Named vector of initial conditions
}
\description{
Generate initial conditions for measles from given proportion
of susceptibles
}
\author{
Sebastian Funk
}

