% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cpx_ms_age.R
\name{ms.prior}
\alias{ms.prior}
\title{Evaluate parameters under a prior distribution for measles}
\usage{
ms.prior(parameters)
}
\arguments{
\item{parameters}{}
}
\value{
prior probability density
}
\description{
Evaluate parameters under a prior distribution for measles
}
\author{
Sebastian Funk
}

