% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pad.zeroes}
\alias{pad.zeroes}
\title{Format a number padded with zeroes}
\usage{
pad.zeroes(no, max)
}
\arguments{
\item{no}{the number to print}

\item{max}{the maximal number (which determines the numbef of zeroes}
}
\value{
padded character
}
\description{
Format a number padded with zeroes
}
\author{
Sebastian Funk
}

