% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpx_ms_age.R
\name{piecewise.linear}
\alias{piecewise.linear}
\title{Generate a piecewise linear trajectory}
\usage{
piecewise.linear(mixing.values, change.times, n)
}
\arguments{
\item{mixing.values}{the mixing values, including initial and final values (so this has to have length two more than \code{change.times}.}

\item{change.times}{the time points at which the linear trajectory changes}

\item{n}{the total number of time points to generate}
}
\value{
the piecewise linear trajectory
}
\description{
Generate a piecewise linear trajectory
}
\author{
Sebastian Funk
}

