% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cpx_ms_age.R
\name{plot.model.data}
\alias{plot.model.data}
\title{Plot a model and data}
\usage{
\method{plot}{model.data}(parameters, data, init, time.column = "time",
  sample.obs = NULL, traj = FALSE)
}
\arguments{
\item{parameters}{model parameters}

\item{data}{data to plot}

\item{init}{initial conditions}

\item{time.column}{time column in the data}

\item{sample.obs}{if given, this well be used to sample observations}
}
\value{
plot
}
\description{
Plot a model and data
}
\author{
Sebastian Funk
}

