% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cpx_ms_age.R
\name{plot.sampled.run}
\alias{plot.sampled.run}
\title{Plot a sampled run and data}
\usage{
plot.sampled.run(theta, data, init.sample, time.column = "time",
  fixed.parameters = NULL, sample.obs = NULL, ...)
}
\arguments{
\item{theta}{parameter vector}

\item{data}{The data to plot}

\item{init.sample}{function to sample initial conditions}

\item{time.column}{time column the data}

\item{fixed.parameters}{fixed parameters}

\item{sample.obs}{sample observation function}

\item{...}{further parameters for \code{plot.model.data}}
}
\value{
plot
}
\description{
Plot a sampled run and data
}
\author{
Sebastian Funk
}

