% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_mixing_vacc_ages}
\alias{plot_mixing_vacc_ages}
\title{Plot age distributions for measles and chickenpox according to
different mixing assumptions}
\usage{
plot_mixing_vacc_ages(ext = NA, classic.life.expectancy = 80,
  kids.mult = 5, max.age = 99)
}
\arguments{
\item{ext}{plot extensions}

\item{classic.life.expectancy}{life expectancy in the "classic" model}

\item{kids.mult}{mixing multiplier for children}

\item{max.age}{maximum age}
}
\description{
Plot age distributions for measles and chickenpox according to
different mixing assumptions
}
\author{
Sebastian Funk
}

