% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plotChickenpoxEWBeforeAfterRatios}
\alias{plotChickenpoxEWBeforeAfterRatios}
\title{Plot England & Wales chickenpox before/after ratios}
\usage{
plotChickenpoxEWBeforeAfterRatios(ext = NA, lines = T)
}
\arguments{
\item{ext}{plot extensions}
}
\description{
Plot England & Wales chickenpox before/after ratios
}
\author{
Sebastian Funk
}

