% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plotChickenpoxEWData}
\alias{plotChickenpoxEWData}
\title{Plot England & Wales chickenpox data}
\usage{
plotChickenpoxEWData(ext = NA)
}
\arguments{
\item{ext}{plot extensions}
}
\description{
Plot England & Wales chickenpox data
}
\author{
Sebastian Funk
}

