% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plotChickenpoxSerology}
\alias{plotChickenpoxSerology}
\title{Plot chickenpox serology}
\usage{
plotChickenpoxSerology(ext = NA)
}
\arguments{
\item{ext}{plot extensions}
}
\description{
Plot chickenpox serology
}
\author{
Sebastian Funk
}

