% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plotDemographicEWData}
\alias{plotDemographicEWData}
\title{Plot England & Wales demographic data}
\usage{
plotDemographicEWData(ext = NA)
}
\arguments{
\item{ext}{plot extensions}
}
\description{
Plot England & Wales demographic data
}
\author{
Sebastian Funk
}

