% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plotMeaslesEuropeanMaps}
\alias{plotMeaslesEuropeanMaps}
\title{Plot European measles maps}
\usage{
plotMeaslesEuropeanMaps(ext)
}
\arguments{
\item{ext}{plot extensions}
}
\description{
Plot European measles maps
}
\author{
Sebastian Funk
}

