% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plotMeaslesEWConfData}
\alias{plotMeaslesEWConfData}
\title{Plot England & Wales measles confirmation data}
\usage{
plotMeaslesEWConfData(ext = NA)
}
\arguments{
\item{ext}{plot extensions}
}
\description{
Plot England & Wales measles confirmation data
}
\author{
Sebastian Funk
}

