% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plotMeaslesEWData}
\alias{plotMeaslesEWData}
\title{Plot England & Wales measles data}
\usage{
plotMeaslesEWData(ext = NA)
}
\arguments{
\item{ext}{plot extensions}
}
\description{
Plot England & Wales measles data
}
\author{
Sebastian Funk
}

