% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plotMeaslesSerology}
\alias{plotMeaslesSerology}
\title{Plot measles serology}
\usage{
plotMeaslesSerology(ext = NA)
}
\arguments{
\item{ext}{plot extensions}
}
\description{
Plot measles serology
}
\author{
Sebastian Funk
}

