% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plotMeaslesWHOData}
\alias{plotMeaslesWHOData}
\title{Plot WHO measles data}
\usage{
plotMeaslesWHOData(ext = NA)
}
\arguments{
\item{ext}{plot extensions}
}
\description{
Plot WHO measles data
}
\author{
Sebastian Funk
}

