% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plotMSCPXEWData}
\alias{plotMSCPXEWData}
\title{Plot England and Wales measles and chicknepox data}
\usage{
plotMSCPXEWData(ext = NA)
}
\arguments{
\item{ext}{plot extensions}
}
\description{
Plot England and Wales measles and chicknepox data
}
\author{
Sebastian Funk
}

