% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{reduce.agegroups}
\alias{reduce.agegroups}
\title{Reduce the number of age groups given a broader set of limtis}
\usage{
reduce.agegroups(x, limits)
}
\arguments{
\item{x}{vector of limits}

\item{limits}{new limits}
}
\value{
vector with the new agegroups
}
\description{
Operates on lower limits
}

