% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seir.R
\name{runSIR}
\alias{runSIR}
\title{Runs an SIR model}
\usage{
runSIR(parameters, init, times, age.labels = NULL, thickening = 1)
}
\arguments{
\item{parameters}{SIR parmaeters}

\item{init}{initial state}

\item{times}{times at which to produce the trajectory}

\item{age.labels}{labels of age groups}

\item{thickening}{whether to thicken the data (use finer integration steps)}

\item{compiled}{whether to use the compiled version of the SIR model}

\item{jacobian}{whether to use the Jacobian}
}
\value{
The initial state and trajectory
}
\description{
Runs an SIR model and returns the trajectory and initial state.
}
\author{
Sebastian Funk
}

