% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polymod.R
\name{sample.contact.matrix}
\alias{sample.contact.matrix}
\title{Samples contact patterns to get an age-based contact matrix using
weights following Baguelin et al. (2013), but using a bootstrap}
\usage{
sample.contact.matrix(participants, contacts, ages, part.age.column,
  contact.age.column = "cnt_age_mean", id.column = "global_id",
  dayofweek.column = "dayofweek", add.weights = c(), symmetry = FALSE)
}
\arguments{
\item{participants}{participant data}

\item{contacts}{contact data}

\item{ages}{age groups}

\item{part.age.column}{column indicating age in participant data}

\item{contact.age.column}{column indicating age in contact data}

\item{id.column}{column to match participants with contacts}

\item{dayofweek.column}{column indicating the day of the week}

\item{add.weights}{additional weight columns (e.g., minutes etc}

\item{symmetry}{make contact matrix symmetric}
}
\value{
age-based mixing matrix
}
\description{
restriction: largest agegroup must not be larger than 90
}
\author{
Sebastian Funk
}

