% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polymod.R
\name{sample.contacts.and.matrix}
\alias{sample.contacts.and.matrix}
\title{samples a contact matrix using weights following Baguelin et al. (2013), but
using a bootstrap; first, contacts and ages are sampled}
\usage{
sample.contacts.and.matrix(n = 1, ...)
}
\arguments{
\item{n}{number of samples}

\item{...}{parameters to pass to sample.contact.matrix}
}
\value{
contact matrix (means and, if n > 1, standard deviation)
}
\description{
restriction: largest agegroup must not be larger than 90
##' @param n number of matrices to sample; if > 1, standard deviation
is reported as well as means
}
\author{
Sebastian Funk
}

