% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpx_ms_age.R
\name{sample.cpx.prior}
\alias{sample.cpx.prior}
\title{Sample parameters from a prior distribution for chickenpox}
\usage{
sample.cpx.prior(fixed.parameters = list())
}
\arguments{
\item{fixed.parameters}{list fixed parameters (won't be sampled)}
}
\value{
sampled parameters
}
\description{
Sample parameters from a prior distribution for chickenpox
}
\author{
Sebastian Funk
}

