% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpx_ms_age.R
\name{sample.init}
\alias{sample.init}
\title{Randomly sample initial conditions for measles or chickenpox on the basis of serology}
\usage{
sample.init(parameters, infection = c("measles", "chickenpox"))
}
\arguments{
\item{parameters}{given (maternal immunity and mixing) parameters}

\item{infection}{"measles" or "chickenpox"}
}
\value{
proportion susceptible in each age group
}
\description{
Randomly sample initial conditions for measles or chickenpox on the basis of serology
}
\author{
Sebastian Funk
}

