% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpx_ms_age.R
\name{sample.ms.init}
\alias{sample.ms.init}
\title{Randomly sample initial conditions for measles on the basis of serology}
\usage{
sample.ms.init(parameters)
}
\arguments{
\item{parameters}{given (maternal immunity and mixing) parameters}
}
\value{
proportion susceptible in each age group
}
\description{
Randomly sample initial conditions for measles on the basis of serology
}
\author{
Sebastian Funk
}

