% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cpx_ms_age.R
\name{sample.ms.prior}
\alias{sample.ms.prior}
\title{Sample parameters from a prior distribution for measles}
\usage{
sample.ms.prior(fixed.parameters = list())
}
\arguments{
\item{mixing}{given mixing matrix, if this is not to sample}

\item{R0}{given R0 value, if this is not to sample}
}
\value{
prior parameters
}
\description{
Sample parameters from a prior distribution for measles
}
\author{
Sebastian Funk
}

