% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpx_ms_age.R
\name{sample.posterior}
\alias{sample.posterior}
\title{Take random samples from the posterior}
\usage{
sample.posterior(sample.prior, eval.prior, gen.init, data, test.init = NULL,
  fixed.parameters = list(), likelihood.arguments = list(),
  mcmc.arguments = list(), time.column = "time", tmax = NULL,
  thickening = 1, interpolate = T, nSamples = 1, update.init = F)
}
\arguments{
\item{sample.prior}{a function to sample from prior}

\item{eval.prior}{a function to evaluate the prior}

\item{gen.init}{a function to generate the initial conditions}

\item{data}{the data to compare the model to}

\item{test.init}{initial conditions (proportoin susceptible in
each each group)}

\item{fixed.parameters}{(list of) fixed parameters of the model}

\item{likelihood.arguments}{(list of) parameters to be passed to
age.likelihood}

\item{mcmc.arguments}{(list of) parameters to be passed to
mcmcMH}

\item{time.column}{the name of the time column in the data}

\item{tmax}{maximum time to run the simulations for}

\item{thickening}{thickening innterval}

\item{interpolate}{whether to intepolate births/deaths etc}

\item{nSamples}{number of samples (if this is 1, the sample will
be from the prior)}

\item{update.init}{whether to update the initial conditions}

\item{return.traj}{whether to return the whol trajectory}
}
\value{
prior, likelihood, posterior, parameters, initial values
}
\description{
Take random samples from the posterior
}
\author{
Sebastian Funk
}

