% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpx_ms_age.R
\name{sample.prior}
\alias{sample.prior}
\title{Sample parameters from a prior distribution for measles or chickenpox}
\usage{
sample.prior(fixed.parameters = list(), infection = c("measles",
  "chickenpox"))
}
\arguments{
\item{infection}{"measles" or "chickenpox"}

\item{mixing}{given mixing matrix, if this is not to sample}
}
\value{
prior parameters
}
\description{
Sample parameters from a prior distribution for measles or chickenpox
}
\author{
Sebastian Funk
}

