% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seir.R
\name{seir.mcmc}
\alias{seir.mcmc}
\title{Run MCMC on the SEIR model}
\usage{
seir.mcmc(target, init.theta, proposal.sd = NULL, n.iterations,
  print.info.every = n.iterations/100, acceptance.rate.window = NULL,
  verbose = FALSE, max.scaling.sd = 50, adapt.size.start = NULL,
  adapt.size.cooling = 0.99, adapt.shape.start = NULL)
}
\arguments{
\item{target}{target function (takes one parameter, a numeric vector)}

\item{init.theta}{initial value of the target function}

\item{proposal.sd}{standard deviation of the proposal kernel}

\item{n.iterations}{number of iterations to run the MCMC for}

\item{print.info.every}{print info every n steps}

\item{acceptance.rate.window}{window to calculate acceptance rate}

\item{verbose}{if TRUE, print verbose output}

\item{max.scaling.sd}{maximal scaling of the standard dveiation in the adaptive MCMC}

\item{adapt.size.start}{after how many iterations to start adapting the size}

\item{adapt.size.cooling}{cooling schedule of size adaptation}

\item{adapt.shape.start}{after how many samples to start adapting the shape}
}
\value{
the trace
}
\description{
Run MCMC on the SEIR model
}
\author{
Anton Camacho, Sebastian Funk
}

