% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seir.R
\name{seir.stochastic}
\alias{seir.stochastic}
\title{calculates a stochastic trajectory of the S(E)IR model}
\usage{
seir.stochastic(parameters, init, seed = NA, transient = 0, tmax = 2128)
}
\arguments{
\item{parameters}{parameters of the S(E)IR mode}

\item{init}{initial state}

\item{seed}{random seed. If set to NA, will use R's current seed}

\item{transient}{transient time. If this is >0, the model will be
run deterministically for the time specified here before running
the stochastic model. Only the outcome of the stochastic model is
recorded}

\item{tmax}{the maximum time to which to run the model (form 0)}

\item{init.params}{whether the initial conditions are passed as parameters}
}
\value{
a stochastic trajectory of the model.
}
\description{
This is being used by \code{\link{runSIR}}.
}
\author{
Sebastian Funk.
}

