% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{year.week.to.date}
\alias{year.week.to.date}
\title{Convert year and week to date.}
\usage{
year.week.to.date(year, week)
}
\arguments{
\item{year}{Year(s)}

\item{week}{Week(s)}
}
\value{
Date(s)
}
\description{
Mostly used to create a date field for ggplot
}

