% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cpx_ms_age.R
\name{cpx.prior}
\alias{cpx.prior}
\title{Evaluate parameters under a prior distribution for chickenpox}
\usage{
cpx.prior(parameters)
}
\arguments{
\item{parameters}{}
}
\value{
prior probability density
}
\description{
Evaluate parameters under a prior distribution for chickenpox
}
\author{
Sebastian Funk
}

