% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_data.R
\name{loadChickenpoxSerology}
\alias{loadChickenpoxSerology}
\title{Load Chickenpox serology data}
\usage{
loadChickenpoxSerology(path)
}
\arguments{
\item{path}{Path in which to find the raw data}
}
\description{
Load Chickenpox serology data
}
\author{
Sebastian Funk
}

