% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{loadMeaslesEWAge}
\alias{loadMeaslesEWAge}
\title{Load age-specific measles England & Wales data}
\usage{
loadMeaslesEWAge(path)
}
\arguments{
\item{path}{Path in which to find the raw data}
}
\description{
Load age-specific measles England & Wales data
}
\author{
Sebastian Funk
}

