% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{reduce.agegroups}
\alias{reduce.agegroups}
\title{Reduce the number of age groups given a broader set of limtis}
\usage{
reduce.agegroups(x, limits, column = "lower.age.limit")
}
\arguments{
\item{x}{data table in which to reduce age groups}

\item{limits}{new limits}

\item{column}{column in x that denotes age}
}
\value{
data table with updated age groups
}
\description{
Operates on lower limits
}

