% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polymod.R
\name{sample.contact.matrix}
\alias{sample.contact.matrix}
\title{Samples contact patterns to get an age-based contact matrix using
weights following Baguelin et al. (2013), but using a bootstrap}
\usage{
sample.contact.matrix(participants, contacts, ages,
  part.age.column = "participant_age", contact.age.column = "contact_age",
  id.column = "global_id", add.weights = c(), symmetry = F)
}
\arguments{
\item{participants}{participant data}

\item{contacts}{contact data}

\item{ages}{age groups}

\item{part.age.column}{column indicating age in participant data}

\item{contact.age.column}{column indicating age in contact data}

\item{id.column}{column to match participants with contacts}

\item{add.weights}{additional weight columns (e.g., minutes etc}

\item{symmetry}{make contact matrix symmetric}
}
\value{
age-based mixing matrix
}
\description{
restriction: largest agegroup must not be larger than 90
}
\author{
Sebastian Funk
}

