% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polymod.R
\name{sample.polymod}
\alias{sample.polymod}
\title{samples polymod using weights following Baguelin et al. (2013), but
using a bootstrap; first, contacts and ages are sampled}
\usage{
sample.polymod(age.limits, countries = NULL, mixing.pop = NULL,
  normalise = FALSE, ...)
}
\arguments{
\item{age.limits}{Lower limits of the age groups}

\item{countries}{limit to one or more countries}

\item{mixing.pop}{mixing population -- either a data frame with columns lower.age.limit and population, or a character vector giving the name(s) to use with the 2013 WHO population}

\item{normalise}{whether to normalise to eigenvalue 1}

\item{...}{further parameters are passed to
\code{sample.contacts.and.matrix}}
}
\value{
sampled contact matrix as return by \code{sample.contacts.and.matrix}
}
\description{
restriction: largest agegroup must not be larger than 90
}
\author{
Sebastian Funk
}

