% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polymod.R
\name{scan.age.dist}
\alias{scan.age.dist}
\title{Scan a range of q values}
\usage{
scan.age.dist(participants, contacts, ages, child.mixing = NA)
}
\arguments{
\item{participants}{participants}

\item{contacts}{contacts}

\item{ages}{ages}

\item{child.mixing}{mixing between under 5 year olds}
}
\value{
a (ggplot) plot with the results
}
\description{
Plots the results
}
\author{
Sebastian Funk
}

